kill super admin, make admin per module
