+++ facebook/modules/profile/user/admin.py
@@ -1,4 +1,3 @@
-from django.conf import settings
 from django.contrib import admin
 
 from facebook.modules.base import AdminBase
@@ -10,5 +9,5 @@ class UserAdmin(AdminBase):
     readonly_fields = ('friends', '_name', '_first_name', '_last_name', '_link', '_birthday', '_email', '_location', '_gender', '_graph')
     search_fields = ('id', '_name')
 
-if not hasattr(settings, 'FACEBOOK_ADMIN'):
-    admin.site.register(User, UserAdmin)
+
+admin.site.register(User, UserAdmin)
